this built, didn't run

+++ Moments/app/src/main/java/moments/moments/MomentPromptActivity.java
@@ -7,6 +7,7 @@ import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.database.Cursor;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
@@ -107,7 +108,7 @@ public class MomentPromptActivity extends Activity implements TextInputDialog.Te
 
 		Cursor allNotes = mDbHelper.fetchAllNotes();
 		int noteCount = allNotes.getCount();
-		int randIndex = Math.floor(Math.random() * noteCount);
+		int randIndex = (int) Math.floor(Math.random() * noteCount);
         allNotes.moveToPosition(randIndex);
 
         Type type;
